+++ test/runtests.jl
@@ -1,5 +1,7 @@
 using Test
 
+include("test_utils.jl")
+
 @testset "BigFloatLinearAlgebra" begin
     include("ownership.jl")
     include("level1.jl")